Fix for last incorrect commit
git-svn-id: svn://dyne.org/rastasoft/freej/freej@204 383723c8-4afa-0310-b8a8-b1afb83214fc

+++ src/include/context.h
@@ -70,7 +70,7 @@ class Context {
 
   bool init(int wx, int hx);
   void close();
-  void cafudda(unsigned int secs);
+  void cafudda(double secs);
 
   /* this returns the address of selected coords to video memory */
   void *coords(int x, int y) { return screen->coords(x,y); };
